+++ tests/collector_add_pools.rs
@@ -277,6 +277,7 @@ async fn test_add_pools_http() -> Result<()> {
             use_websocket: false,
             websocket_urls: vec![],
             wait_time: 2_000, // poll every 2 s
+            refetch_algebra_fee: false,
         },
         Arc::clone(&registry),
         None,
@@ -398,6 +399,7 @@ async fn test_add_pools_ws() -> Result<()> {
             use_websocket: true,
             websocket_urls: WS_RPCS.iter().map(|s| s.to_string()).collect(),
             wait_time: 0,
+            refetch_algebra_fee: false,
         },
         Arc::clone(&registry),
         None,